HIRO/requirements.txt
opencv-python
pillow
speechrecognition
pyttsx3
moondream
groq
